Rename rule to 'no-mime-sniff'
