+++ eslint-bridge/src/rules/main.ts
@@ -99,6 +99,7 @@ import { rule as noInconsistentReturns } from './no-inconsistent-returns';
 import { rule as noInvalidAwait } from './no-invalid-await';
 import { rule as noInvariantReturns } from './no-invariant-returns';
 import { rule as noLabels } from './no-labels';
+import { rule as noMimeSniff } from './no-mime-sniff';
 import { rule as noMisleadingArrayReverse } from './no-misleading-array-reverse';
 import { rule as noNestedAssignment } from './no-nested-assignment';
 import { rule as noNestedConditional } from './no-nested-conditional';
@@ -112,7 +113,6 @@ import { rule as noRedundantOptional } from './no-redundant-optional';
 import { rule as noRedundantParentheses } from './no-redundant-parentheses';
 import { rule as noRequireOrDefine } from './no-require-or-define';
 import { rule as noReturnTypeAny } from './no-return-type-any';
-import { rule as noSniff } from './no-sniff';
 import { rule as noTab } from './no-tab';
 import { rule as noTryPromise } from './no-try-promise';
 import { rule as noUndefinedArgument } from './no-undefined-argument';
@@ -247,6 +247,7 @@ ruleModules['no-inconsistent-returns'] = noInconsistentReturns;
 ruleModules['no-invalid-await'] = noInvalidAwait;
 ruleModules['no-invariant-returns'] = noInvariantReturns;
 ruleModules['no-labels'] = noLabels;
+ruleModules['no-mime-sniff'] = noMimeSniff;
 ruleModules['no-misleading-array-reverse'] = noMisleadingArrayReverse;
 ruleModules['no-nested-assignment'] = noNestedAssignment;
 ruleModules['no-nested-conditional'] = noNestedConditional;
@@ -260,7 +261,6 @@ ruleModules['no-redundant-optional'] = noRedundantOptional;
 ruleModules['no-redundant-parentheses'] = noRedundantParentheses;
 ruleModules['no-require-or-define'] = noRequireOrDefine;
 ruleModules['no-return-type-any'] = noReturnTypeAny;
-ruleModules['no-sniff'] = noSniff;
 ruleModules['no-tab'] = noTab;
 ruleModules['no-try-promise'] = noTryPromise;
 ruleModules['no-undefined-argument'] = noUndefinedArgument;
