+++ eslint-bridge/src/rules/no-mime-sniff.ts
