+++ javascript-checks/src/main/java/org/sonar/javascript/checks/CheckList.java
@@ -219,12 +219,12 @@ public final class CheckList {
       NoInMisuseCheck.class,
       NoSparseArraysCheck.class,
       NotStoredSelectionCheck.class,
+      NoMimeSniffCheck.class,
       NoNonNullAssertionCheck.class,
       NoNestedTemplateLiteralsCheck.class,
       NoRedundantOptionalCheck.class,
       NoReturnAwaitCheck.class,
       NoReturnTypeAnyCheck.class,
-      NoSniffCheck.class,
       NoUnnecessaryTypeAssertionCheck.class,
       NoUselessCatchCheck.class,
       NoWeakCipherCheck.class,
